+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/TitleDemo/TitleDemo.tsx
@@ -1,16 +1,16 @@
-import { Title,  TitleProps, TitleLevel } from '@patternfly/react-core';
+import { Title,  TitleLevel, TitleProps } from '@patternfly/react-core';
 import React, { Component } from 'react';
 
-export class TitleDemo extends Component {
-  myTitleProps: TitleProps = {
-    size: 'md',
-    headingLevel: TitleLevel.h1
-  };
+const myProps: TitleProps = {
+  size: 'md',
+  headingLevel: TitleLevel.h1
+};
 
+export class TitleDemo extends Component {
   render() {
     return (
     <React.Fragment>
-      <Title size={this.myTitleProps.size} headingLevel={this.myTitleProps.headingLevel}>
+      <Title size={myProps.size} headingLevel={myProps.headingLevel}>
         Integration Demo App
       </Title>
     </React.Fragment>